+++ domains/certificates/RegistryConnector/Worker/PeriodExtension.cs
@@ -1,0 +1,15 @@
+using System;
+using DomainCertificate.ValueObjects;
+using ProjectOrigin.Electricity.Client.Models;
+
+namespace RegistryConnector.Worker
+{
+    public static class PeriodExtension
+    {
+        public static DateInterval ToDateInterval(this Period period) =>
+            new(
+                DateTimeOffset.FromUnixTimeSeconds(period.DateFrom),
+                DateTimeOffset.FromUnixTimeSeconds(period.DateTo)
+            );
+    }
+}
