+++ domains/certificates/Shared/DomainCertificate/DomainCertificate.csproj
@@ -6,8 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="ProjectOrigin.Electricity.Client" Version="0.1.0-alpha.19" />
-  </ItemGroup>
-
 </Project>
